requirements.txt
ddgs>=4.0.0
groq==0.4.2
google-generativeai>=0.8.0
python-dotenv==1.0.0
requests==2.31.0
arxiv==1.4.7
